docs: add swagger module
