+++ src/presentation/app.module.ts
@@ -4,6 +4,8 @@ import { envSchema } from '../infrastructure/env/env';
 import { AuthModule } from '@/infrastructure/auth/auth.module';
 import { HttpModule } from './controllers/http.module';
 import { EnvModule } from '@/infrastructure/env/env.module';
+import { SwaggerModule } from '@nestjs/swagger';
+import { SwaggerService } from './docs/swagger.service';
 
 @Module({
   imports: [
@@ -14,6 +16,8 @@ import { EnvModule } from '@/infrastructure/env/env.module';
     AuthModule,
     HttpModule,
     EnvModule,
+    SwaggerModule,
   ],
+  providers: [SwaggerService],
 })
 export class AppModule {}
