+++ src/presentation/docs/swagger.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { SwaggerService } from './swagger.service';
+
+@Module({
+  providers: [SwaggerService],
+  exports: [SwaggerService],
+})
+export class SwaggerModule {}
